// src/constants/enums/messages.ts
export enum Messages {
    SERVER_ERROR = "Internal server error",
    INVALID_BODY = "Request body does not contain required fields",
    INVALID_ENDPOINT = "Invalid endpoint",
    INVALID_USER_ID = "User id is invalid",
    NOT_FOUND_USER = "User doesn't exist",
    USER_DELETED = "User deleted",
}